Wired upp all components except rank fusion.
