+++ lib/getResults.js
@@ -8,7 +8,7 @@ function setVariables(pipeline,searchIndex,vectorIndex,searchCollection){
             }else if('$vectorSearch' in stage){
                 newPipeline.push({...stage,$vectorSearch:{...stage.$vectorSearch, index:vectorIndex}})
             }else if('$unionWith' in stage){
-                newPipeline.push({...stage,$unionWith:{...stage.$unionWith,coll:searchCollection,pipeline:setVariables(stage.$unionWith.pipeline)}})
+                newPipeline.push({...stage,$unionWith:{...stage.$unionWith,coll:searchCollection,pipeline:setVariables(stage.$unionWith.pipeline,searchIndex,vectorIndex,searchCollection)}})
             }else{
                 newPipeline.push(stage);
             }
